[openmct-YAMCS] TelemetryTable does not show limit styles.
