+++ src/providers/realtime-telemetry-provider.js
@@ -30,106 +30,6 @@ import {
 
 const WS_IDLE_INTERVAL_MS = 10000;
 const FALLBACK_AND_WAIT_MS = [1000, 5000, 5000, 10000, 10000, 30000];
-
-/* CSS classes for Yamcs parameter monitoring result values. */
-const MONITORING_RESULT_CSS = {
-    'WATCH': 'is-limit--yellow',
-    'WARNING': 'is-limit--yellow',
-    'DISTRESS': 'is-limit--red',
-    'CRITICAL': 'is-limit--red',
-    'SEVERE': 'is-limit--red'
-};
-
-/* CSS classes for Yamcs range condition values. */
-const RANGE_CONDITION_CSS = {
-    'LOW': 'is-limit--lwr',
-    'HIGH': 'is-limit--upr'
-};
-
-
-/*
- * Implements an OpenMCT limit evaluator that uses limit information passed
- * from either the historical or realtime telemetry source to create limit
- * information for OpenMCT.
- */
-class LimitEvaluator {
-
-    /*
-     * Evaluates a telemetry point for limit violations.
-     *
-     * Parameters:
-     *     datum: the telemetry point data from the historical or realtime
-     *            plugin
-     *     valueMetadata: metadata about the telemetry point, not used
-     *
-     * Returns:
-     *     an object with CSS class information, a violation name, and
-     *     range information, if there is a limit violates, or nothing,
-     *     if the value is within limits.
-     *
-     * The datum parameter may include the following information:
-     *     monitoringResult: the Yamcs limit monitoring result, if any
-     *     rangeCondition: the Yamcs range condition (LOW/HIGH), if any
-     *     alarmRange: an array of alarm ranges for different monitoring
-     *         results, or omitted, if no alarm ranges are defined
-     */
-    evaluate(datum, valueMetadata) {
-        if (datum.monitoringResult
-            && datum.monitoringResult in MONITORING_RESULT_CSS) {
-            let evaluationResult = {
-                cssClass: MONITORING_RESULT_CSS[datum.monitoringResult],
-                name: datum.monitoringResult
-            };
-            if (datum.rangeCondition
-                && datum.rangeCondition in RANGE_CONDITION_CSS) {
-                evaluationResult.cssClass += ' ';
-                evaluationResult.cssClass += RANGE_CONDITION_CSS[datum.rangeCondition];
-                this.addLimitRange(datum, datum.monitoringResult, evaluationResult);
-            }
-            return evaluationResult;
-        }
-    }
-
-    /*
-     * Adds limit range information to an object based on the monitoring
-     * result.
-     *
-     * Parameters:
-     *     datum: the telemetry point data from the historical or realtime
-     *            plugin
-     *     result: the monitoring result information from Yamcs
-     *     evaluationResult: the object that is to be the result of the
-     *                       limit evaluation
-     */
-    addLimitRange(datum, result, evaluationResult) {
-        let range = this.findAlarmRange(datum, result);
-        if (range === undefined) {
-            return;
-        }
-        if (range.minInclusive) {
-            evaluationResult.low = range.minInclusive;
-        }
-        if (range.minExclusive) {
-            evaluationResult.low = range.minExclusive;
-        }
-        if (range.maxInclusive) {
-            evaluationResult.high = range.maxInclusive;
-        }
-        if (range.maxExclusive) {
-            evaluationResult.high = range.maxExclusive;
-        }
-    }
-
-    /*
-     * Finds the appropriate limit range for a monitoring results.
-     */
-    findAlarmRange(datum, result) {
-        if (datum.alarmRange !== undefined) {
-            return datum.alarmRange.find(range => range.level === result);
-        }
-    }
-}
-
 export default class RealtimeTelemetryProvider {
     constructor(url ,instance) {
         this.url = url;
@@ -153,10 +53,6 @@ export default class RealtimeTelemetryProvider {
 
     supportsSubscribe(domainObject) {
         return this.supportedTypes[domainObject.type];
-    }
-
-    supportsLimits(domainObject) {
-        return domainObject.type.startsWith('yamcs.');
     }
 
     subscribe(domainObject, callback) {
@@ -191,10 +87,6 @@ export default class RealtimeTelemetryProvider {
             this.eventUnsubscribe();
             delete this.listener[id];
         };
-    }
-
-    getLimitEvaluator(domainObject) {
-        return new LimitEvaluator();
     }
 
     resubscribeToAll() {
